refactor(reco-core): zero-copy projections - the document is the engine
impl Projection lands directly on the topology param structs; the
LShapeProjection/CylindricalProjection unit structs and for_topology
are deleted. Topology::projection() hands out a borrow of the active
variant, so there is no second object to build or keep in sync and the
render-time topology downcasts (.l_shape()/.cylinder().expect) are
unrepresentable: each impl reads self.

Executors stop owning a projection: CpuExecutor::new loses its
projection parameter and reads the document per call (a calibration
swap switches projection automatically); GpuExecutor keeps only the
injected out-of-tree override, logged as such. The redundant lens-count
check in CpuExecutor::new was already covered by validate().

Projection::camera_count widens u8 -> usize to match the lens-count
comparisons it exists for.

+++ crates/reco-core/src/stitch/cpu.rs
@@ -121,7 +121,7 @@ pub(crate) fn stitch_rgba_yuv420p(
 /// The number of input frames must match what the projection consumes;
 /// a mismatch would silently sample the wrong camera.
 fn check_camera_count(projection: &dyn Projection, planes: usize) -> Result<(), StitchError> {
-    if planes != usize::from(projection.camera_count()) {
+    if planes != projection.camera_count() {
         return Err(StitchError::InvalidConfig(format!(
             "projection '{}' consumes {} camera(s) but {planes} frame(s) were supplied",
             projection.name(),
